refactor: change float type to double for more precision

+++ src/main/java/Level1/Sale.java
@@ -1,12 +1,11 @@
 package Level1;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
 public class Sale {
     private List<Product> products = new ArrayList<>();
-    private BigDecimal totalPrice;
+    private double totalPrice;
 
     public void calculateTotalPrice() {
         try {
